fin de journée 29/07

+++ src/Controller/Forum/ForumController.php
@@ -1,5 +1,7 @@
 <?php
 
+// src/Controller/Forum/ForumController.php
+
 namespace App\Controller\Forum;
 
 use App\Entity\Category;
@@ -20,7 +22,6 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
-
 class ForumController extends AbstractController
 {
     private EntityManagerInterface $entityManager;
@@ -35,10 +36,7 @@ class ForumController extends AbstractController
     #[Route('/forum', name: 'app_forum', methods: ['GET'])]
     public function index(CardRepository $cardRepository, CategoryRepository $categoryRepository): Response
     {
-        // Récupérer toutes les catégories
         $categories = $categoryRepository->findAll();
-        
-        // Récupérer uniquement les cartes avec les IDs 34, 35, 36
         $cardIds = [34, 35, 36];
         $cards = $cardRepository->findBy(['id' => $cardIds]);
 
@@ -54,25 +52,30 @@ class ForumController extends AbstractController
         $topics = $category->getTopics();
 
         $topic = new Topic();
-        $topic->setCategory($category);
-        $topic->setUser($this->getUser()); // Assuming user is authenticated
-        $topic->setCreationDate(new \DateTime()); // Set the creation date
+        if ($this->getUser()) {
+            $topic->setCategory($category);
+            $topic->setUser($this->getUser());
+            $topic->setCreationDate(new \DateTime());
 
-        $form = $this->createForm(TopicType::class, $topic);
-        $form->handleRequest($request);
+            $form = $this->createForm(TopicType::class, $topic);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->entityManager->persist($topic);
-            $this->entityManager->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $this->entityManager->persist($topic);
+                $this->entityManager->flush();
 
-            return $this->redirectToRoute('forum_topic_show', ['id' => $topic->getId()]);
+                return $this->redirectToRoute('forum_topic_show', ['id' => $topic->getId()]);
+            }
+
+            return $this->render('forum/category_show.html.twig', [
+                'category' => $category,
+                'topics' => $topics,
+                'form' => $form->createView(),
+            ]);
+        } else {
+            // Rediriger vers la page de connexion
+            return $this->redirectToRoute('app_login');
         }
-
-        return $this->render('forum/category_show.html.twig', [
-            'category' => $category,
-            'topics' => $topics,
-            'form' => $form->createView(),
-        ]);
     }
 
     #[Route('/forum/category/{categoryId}/new-topic', name: 'forum_new_topic', methods: ['GET', 'POST'])]
@@ -80,24 +83,28 @@ class ForumController extends AbstractController
     {
         $category = $this->entityManager->getRepository(Category::class)->find($categoryId);
         $topic = new Topic();
-        $topic->setCategory($category);
-        $topic->setUser($this->getUser()); // Assuming user is authenticated
-        $topic->setCreationDate(new \DateTime()); // Set the creation date
+        if ($this->getUser()) {
+            $topic->setCategory($category);
+            $topic->setUser($this->getUser());
+            $topic->setCreationDate(new \DateTime());
 
-        $form = $this->createForm(TopicType::class, $topic);
-        $form->handleRequest($request);
+            $form = $this->createForm(TopicType::class, $topic);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->entityManager->persist($topic);
-            $this->entityManager->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $this->entityManager->persist($topic);
+                $this->entityManager->flush();
 
-            return $this->redirectToRoute('forum_topic_show', ['id' => $topic->getId()]);
+                return $this->redirectToRoute('forum_topic_show', ['id' => $topic->getId()]);
+            }
+
+            return $this->render('forum/new_topic.html.twig', [
+                'category' => $category,
+                'form' => $form->createView(),
+            ]);
+        } else {
+            return $this->redirectToRoute('app_login');
         }
-
-        return $this->render('forum/new_topic.html.twig', [
-            'category' => $category,
-            'form' => $form->createView(),
-        ]);
     }
 
     #[Route('/forum/topic/{id}', name: 'forum_topic_show', methods: ['GET', 'POST'])]
@@ -113,32 +120,40 @@ class ForumController extends AbstractController
         $posts = $this->entityManager->getRepository(Post::class)->findByTopicPaginated($id, $page);
 
         $post = new Post();
-        $post->setTopic($topic);
-        $post->setUser($this->getUser()); // Assuming user is authenticated
+        if ($this->getUser()) {
+            $post->setTopic($topic);
+            $post->setUser($this->getUser());
 
-        $form = $this->createForm(PostType::class, $post);
-        $form->handleRequest($request);
+            $form = $this->createForm(PostType::class, $post);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->entityManager->persist($post);
-            $this->entityManager->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $this->entityManager->persist($post);
+                $this->entityManager->flush();
 
-            return $this->redirectToRoute('forum_topic_show', ['id' => $topic->getId()]);
+                return $this->redirectToRoute('forum_topic_show', ['id' => $topic->getId()]);
+            }
+
+            return $this->render('forum/topic_show.html.twig', [
+                'topic' => $topic,
+                'posts' => $posts,
+                'form' => $form->createView(),
+            ]);
+        } else {
+            return $this->redirectToRoute('app_login');
         }
-
-        return $this->render('forum/topic_show.html.twig', [
-            'topic' => $topic,
-            'posts' => $posts,
-            'form' => $form->createView(),
-        ]);
     }
 
     #[Route('/forum/post/{id}/report', name: 'forum_post_report', methods: ['GET', 'POST'])]
     public function reportPost(Request $request, Post $post): Response
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('app_login');
+        }
+
         $report = new UserReport();
         $report->setPost($post);
-        $report->setReportingUser($this->getUser()); // Assuming user is authenticated
+        $report->setReportingUser($this->getUser());
         $report->setReportedUser($post->getUser());
 
         $form = $this->createForm(UserReportType::class, $report);
@@ -156,44 +171,40 @@ class ForumController extends AbstractController
             'form' => $form->createView(),
         ]);
     }
-    
+
     #[Route('/forum/post/{id}/edit', name: 'forum_post_edit', methods: ['GET', 'POST'])]
-public function editPost(Request $request, Post $post): Response
-{
-    // Vérifiez si l'utilisateur connecté est l'auteur du post
-    if ($post->getUser() !== $this->getUser()) {
-        throw new AccessDeniedException('You are not allowed to edit this post.');
+    public function editPost(Request $request, Post $post): Response
+    {
+        if (!$this->getUser() || $post->getUser() !== $this->getUser()) {
+            throw new AccessDeniedException('You are not allowed to edit this post.');
+        }
+
+        $form = $this->createForm(PostType::class, $post);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $this->entityManager->flush();
+            return $this->redirectToRoute('forum_topic_show', ['id' => $post->getTopic()->getId()]);
+        }
+
+        return $this->render('forum/edit_post.html.twig', [
+            'post' => $post,
+            'form' => $form->createView(),
+        ]);
     }
 
-    $form = $this->createForm(PostType::class, $post);
-    $form->handleRequest($request);
+    #[Route('/forum/post/{id}/delete', name: 'forum_post_delete', methods: ['POST'])]
+    public function deletePost(Request $request, Post $post): Response
+    {
+        if (!$this->getUser() || $post->getUser() !== $this->getUser()) {
+            throw new AccessDeniedException('You are not allowed to delete this post.');
+        }
 
-    if ($form->isSubmitted() && $form->isValid()) {
-        $this->entityManager->flush();
+        if ($this->isCsrfTokenValid('delete'.$post->getId(), $request->request->get('_token'))) {
+            $this->entityManager->remove($post);
+            $this->entityManager->flush();
+        }
+
         return $this->redirectToRoute('forum_topic_show', ['id' => $post->getTopic()->getId()]);
     }
-
-    return $this->render('forum/edit_post.html.twig', [
-        'post' => $post,
-        'form' => $form->createView(),
-    ]);
-}
-
-#[Route('/forum/post/{id}/delete', name: 'forum_post_delete', methods: ['POST'])]
-public function deletePost(Request $request, Post $post): Response
-{
-    // Vérifiez si l'utilisateur connecté est l'auteur du post
-    if ($post->getUser() !== $this->getUser()) {
-        throw new AccessDeniedException('You are not allowed to delete this post.');
-    }
-
-    // Vérifiez le token CSRF
-    if ($this->isCsrfTokenValid('delete'.$post->getId(), $request->request->get('_token'))) {
-        $this->entityManager->remove($post);
-        $this->entityManager->flush();
-    }
-
-    return $this->redirectToRoute('forum_topic_show', ['id' => $post->getTopic()->getId()]);
-}
-    
 }
